Safari-611.1.10/LayoutTests/crypto/subtle/hmac-generate-key-sign-verify-expected.txt

Test sign/verify with HMAC SHA-1 using a generated key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating key...
Signing...
PASS signature.byteLength is 20
Verifying...
PASS verified is true
PASS successfullyParsed is true

TEST COMPLETE
